+++ src/SMTP2GO/Collections/Collection.php
@@ -2,7 +2,7 @@
 
 namespace SMTP2GO\Collections;
 
-class Collection implements \ArrayAccess, \Iterator
+class Collection implements \ArrayAccess, \Iterator, \Countable
 {
     protected $items;
     protected $position;
@@ -90,5 +90,11 @@ class Collection implements \ArrayAccess, \Iterator
     public function valid(): bool
     {
         return isset($this->items[$this->position]);
+    }
+
+    /** @return int  */
+    public function count(): int
+    {
+        return count($this->items);
     }
 }