+++ src/SMTP2GO/Collections/Mail/AttachmentCollection.php
@@ -8,7 +8,7 @@ use SMTP2GO\Types\Mail\FileAttachment;
 
 class AttachmentCollection extends Collection
 {
-    protected $items;
+    protected $items = [];
 
     public function __construct(array $attachments = [])
     {
@@ -22,7 +22,7 @@ class AttachmentCollection extends Collection
         if (is_a($attachment, Attachment::class) || is_a($attachment, FileAttachment::class)) {
             $this->items[] = $attachment;
         } else {
-            throw new \InvalidArgumentException('This collection expects objects of type ' . Attachment::class, ' but recieved ' . get_class($attachment));
+            throw new \InvalidArgumentException('This collection expects objects of type ' . Attachment::class, ' but received ' . get_class($attachment));
         }
         return $this;
     }