+++ src/SMTP2GO/Service/Mail/Send.php
@@ -2,9 +2,6 @@
 
 namespace SMTP2GO\Service\Mail;
 
-
-use SMTP2GO\Mime\Detector;
-
 use InvalidArgumentException;
 use SMTP2GO\Types\Mail\Address;
 use SMTP2GO\Types\Mail\Attachment;
@@ -14,6 +11,7 @@ use SMTP2GO\Types\Mail\InlineAttachment;
 use SMTP2GO\Collections\Mail\AddressCollection;
 use SMTP2GO\Collections\Mail\AttachmentCollection;
 use SMTP2GO\Types\Mail\CustomHeader;
+use SMTP2GO\Types\Mail\FileAttachment;
 
 /**
  * Constructs the payload for sending email through the SMTP2GO Api
@@ -83,7 +81,7 @@ class Send implements BuildsRequest
     /**
      * The template data to use which is key value pairs of [placeholder => replacement]
      * @link https://app-us.smtp2go.com/settings/templates/
-     * @var array
+     * @var array|null
      */
     protected $template_data;
 
@@ -109,13 +107,26 @@ class Send implements BuildsRequest
     protected $inlines;
 
 
-    /**
+    /**      
      * @var int version
-     * The version parameter specifies which version (structure) to use when generating the email
-     * @see https://apidoc.smtp2go.com/documentation/#/POST%20/email/send
+     * @deprecated 
      * 
      */
     protected $version = 1;
+
+    /**
+     * @var int scheduleAt
+     * A unix timestamp to schedule the email for sending in the future.
+     * 
+     */
+    protected $scheduleAt = null;
+
+
+    /**
+     * If true, the email will be accepted immediately and sent in a background process. Use webhooks if you need information about final delivery to the recipient. This will soon become the default method of sending via API.
+     * @var bool
+     */
+    protected $fastaccept = false;
 
     /**
      * endpoint to send to
@@ -177,10 +188,20 @@ class Send implements BuildsRequest
         $body['template_id'] = $this->template_id ?? null;
         $body['template_data'] = $this->template_data ?? null;
         $body['version'] = $this->version;
+        $body['schedule'] = $this->scheduleAt;
+        $body['fastaccept'] = $this->fastaccept;
 
         return array_filter($body);
     }
 
+    public function scheduleAt(int $timestamp)
+    {
+        if ($timestamp < time() || $timestamp > time() + (3 * 24 * 60 * 60)) {
+            throw new \InvalidArgumentException('The timestamp must be a valid unix timestamp in the future, and no more than 3 days from now.');
+        }
+        $this->scheduleAt = $timestamp;
+        return $this;
+    }
 
 
     public function buildCustomHeaders()
@@ -200,7 +221,7 @@ class Send implements BuildsRequest
      */
     public function buildAttachments(): array
     {
-        if (empty($this->attachments)) {
+        if ($this->attachments->count() === 0) {
             return [];
         }
 
@@ -220,7 +241,7 @@ class Send implements BuildsRequest
      */
     public function buildInlines(): array
     {
-        if (empty($this->inlines)) {
+        if ($this->inlines->count() === 0) {
             return [];
         }
 
@@ -269,9 +290,8 @@ class Send implements BuildsRequest
     /**
      * Add a custom header
      *
-     * @param string $headerName
-     * @param string $headerValue
-     * @return void
+     * @param  CustomHeader $header
+     * @return Send
      */
     public function addCustomHeader(CustomHeader $header): Send
     {
@@ -282,7 +302,7 @@ class Send implements BuildsRequest
     /**
      * Get sender
      *
-     * @return  string
+     * @return string
      */
     public function getSender(): string
     {
@@ -347,7 +367,7 @@ class Send implements BuildsRequest
     /**
      * Set the email message
      *
-     * @param  string  $message  The email message
+     * @param  string  $htmlBody  The email message
      *
      * @return Send
      */
@@ -388,7 +408,7 @@ class Send implements BuildsRequest
      * add multiple addresses of a specified type
      *
      * @param string $addressType either 'to', 'cc', 'bcc'
-     * @param array $addresses the array should contain multiple arrays with 1 or 2 values with email address and optional name
+     * @param AddressCollection $addresses the array should contain multiple arrays with 1 or 2 values with email address and optional name
      * @return Send
      */
     public function addAddresses(string $addressType, AddressCollection $addresses): Send
@@ -404,8 +424,7 @@ class Send implements BuildsRequest
      * add a single addresses of a specified type
      *
      * @param string $addressType either 'to', 'cc', 'bcc'
-     * @param string $email
-     * @param string $name
+     * @param Address $address
      * @return Send
      */
 
@@ -454,9 +473,9 @@ class Send implements BuildsRequest
     /**
      * Get the CC'd recipients. This clears any previously added CC addresses
      *
-     * @return  AddressCollection
+     * @return array
      */
-    public function getCc()
+    public function getCc(): array
     {
         return $this->cc;
     }
@@ -487,11 +506,14 @@ class Send implements BuildsRequest
         return $this->attachments;
     }
 
-    public function addAttachment(Attachment $attachment): Send
+    public function addAttachment($attachment): Send
     {
         if (is_a($attachment, InlineAttachment::class)) {
             $this->inlines[] = $attachment;
-        } else {
+        } elseif (
+            is_a($attachment, FileAttachment::class)
+            || is_a($attachment, Attachment::class)
+        ) {
             $this->attachments[] = $attachment;
         }
         return $this;
@@ -636,6 +658,30 @@ class Send implements BuildsRequest
     public function setVersion(int $version)
     {
         $this->version = $version;
+
+        return $this;
+    }
+
+    /**
+     * 
+     *
+     * @return  bool
+     */
+    public function getFastaccept()
+    {
+        return $this->fastaccept;
+    }
+
+    /**
+     * 
+     *
+     * @param  bool  $fastaccept
+     *
+     * @return  self
+     */
+    public function setFastaccept(bool $fastaccept)
+    {
+        $this->fastaccept = $fastaccept;
 
         return $this;
     }