+++ src/SMTP2GO/Types/Mail/CustomHeader.php
@@ -4,7 +4,7 @@ namespace SMTP2GO\Types\Mail;
 
 class CustomHeader
 {
-    protected $header;
+    protected $header = '';
 
     protected $value = '';
 
@@ -33,7 +33,7 @@ class CustomHeader
      *
      * @return  self
      */
-    public function setValue($value)
+    public function setValue(string $value)
     {
         $this->value = $value;
 
@@ -53,7 +53,7 @@ class CustomHeader
      *
      * @return  self
      */
-    public function setHeader($header)
+    public function setHeader(string $header)
     {
         $this->header = $header;