+++ src/SMTP2GO/Types/Mail/FileAttachment.php
@@ -22,7 +22,7 @@ class FileAttachment
      */
     protected $mimetype;
 
-    public function __construct($body, $filename)
+    public function __construct(string $body, string $filename)
     {
 
         $detector = new Detector;
@@ -53,7 +53,7 @@ class FileAttachment
      *
      * @return  self
      */
-    public function setBody($body)
+    public function setBody(string $body)
     {
         $this->body = $body;
 
@@ -65,7 +65,7 @@ class FileAttachment
      *
      * @return  self
      */
-    public function setFilename($filename)
+    public function setFilename(string $filename)
     {
         $this->filename = $filename;
 
@@ -93,7 +93,7 @@ class FileAttachment
      *
      * @return  self
      */
-    public function setMimetype($mimetype)
+    public function setMimetype(string $mimetype)
     {
         $this->mimetype = $mimetype;