+++ tests/Unit/ServiceTest.php
@@ -1,0 +1,14 @@
+<?php
+
+use SMTP2GO\Service\Service;
+use PHPUnit\Framework\TestCase;
+
+class ServiceTest extends TestCase
+{
+    public function testSettingEndpointWithLeadingSlashStripsSlash()
+    {
+        $service = new Service('/email/send',[]);
+        
+        $this->assertEquals('email/send', $service->getEndpoint());
+    }
+}